+++ clients/spot/src/restapi/models/model_historical_block_trades_response_inner.go
@@ -1,0 +1,336 @@
+/*
+Binance Spot REST API
+
+OpenAPI Specifications for the Binance Spot REST API  API documents:   - [Github rest-api documentation file](https://github.com/binance/binance-spot-api-docs/blob/master/rest-api.md)   - [General API information for rest-api on website](https://developers.binance.com/docs/binance-spot-api-docs/rest-api/general-api-information)
+*/
+
+package models
+
+import (
+	"encoding/json"
+
+	"github.com/binance/binance-connector-go/common/v2/common"
+)
+
+// checks if the HistoricalBlockTradesResponseInner type satisfies the MappedNullable interface at compile time
+var _ common.MappedNullable = &HistoricalBlockTradesResponseInner{}
+
+// HistoricalBlockTradesResponseInner struct for HistoricalBlockTradesResponseInner
+type HistoricalBlockTradesResponseInner struct {
+	Id                   *int64  `json:"id,omitempty"`
+	Price                *string `json:"price,omitempty"`
+	Qty                  *string `json:"qty,omitempty"`
+	QuoteQty             *string `json:"quoteQty,omitempty"`
+	Time                 *int64  `json:"time,omitempty"`
+	IsBuyerMaker         *bool   `json:"isBuyerMaker,omitempty"`
+	AdditionalProperties map[string]interface{}
+}
+
+type _HistoricalBlockTradesResponseInner HistoricalBlockTradesResponseInner
+
+// NewHistoricalBlockTradesResponseInner instantiates a new HistoricalBlockTradesResponseInner object
+// This constructor will assign default values to properties that have it defined,
+// and makes sure properties required by API are set, but the set of arguments
+// will change when the set of required properties is changed
+func NewHistoricalBlockTradesResponseInner() *HistoricalBlockTradesResponseInner {
+	this := HistoricalBlockTradesResponseInner{}
+	return &this
+}
+
+// NewHistoricalBlockTradesResponseInnerWithDefaults instantiates a new HistoricalBlockTradesResponseInner object
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewHistoricalBlockTradesResponseInnerWithDefaults() *HistoricalBlockTradesResponseInner {
+	this := HistoricalBlockTradesResponseInner{}
+	return &this
+}
+
+// GetId returns the Id field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetId() int64 {
+	if o == nil || common.IsNil(o.Id) {
+		var ret int64
+		return ret
+	}
+	return *o.Id
+}
+
+// GetIdOk returns a tuple with the Id field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetIdOk() (*int64, bool) {
+	if o == nil || common.IsNil(o.Id) {
+		return nil, false
+	}
+	return o.Id, true
+}
+
+// HasId returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasId() bool {
+	if o != nil && !common.IsNil(o.Id) {
+		return true
+	}
+
+	return false
+}
+
+// SetId gets a reference to the given int64 and assigns it to the Id field.
+func (o *HistoricalBlockTradesResponseInner) SetId(v int64) {
+	o.Id = &v
+}
+
+// GetPrice returns the Price field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetPrice() string {
+	if o == nil || common.IsNil(o.Price) {
+		var ret string
+		return ret
+	}
+	return *o.Price
+}
+
+// GetPriceOk returns a tuple with the Price field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetPriceOk() (*string, bool) {
+	if o == nil || common.IsNil(o.Price) {
+		return nil, false
+	}
+	return o.Price, true
+}
+
+// HasPrice returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasPrice() bool {
+	if o != nil && !common.IsNil(o.Price) {
+		return true
+	}
+
+	return false
+}
+
+// SetPrice gets a reference to the given string and assigns it to the Price field.
+func (o *HistoricalBlockTradesResponseInner) SetPrice(v string) {
+	o.Price = &v
+}
+
+// GetQty returns the Qty field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetQty() string {
+	if o == nil || common.IsNil(o.Qty) {
+		var ret string
+		return ret
+	}
+	return *o.Qty
+}
+
+// GetQtyOk returns a tuple with the Qty field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetQtyOk() (*string, bool) {
+	if o == nil || common.IsNil(o.Qty) {
+		return nil, false
+	}
+	return o.Qty, true
+}
+
+// HasQty returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasQty() bool {
+	if o != nil && !common.IsNil(o.Qty) {
+		return true
+	}
+
+	return false
+}
+
+// SetQty gets a reference to the given string and assigns it to the Qty field.
+func (o *HistoricalBlockTradesResponseInner) SetQty(v string) {
+	o.Qty = &v
+}
+
+// GetQuoteQty returns the QuoteQty field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetQuoteQty() string {
+	if o == nil || common.IsNil(o.QuoteQty) {
+		var ret string
+		return ret
+	}
+	return *o.QuoteQty
+}
+
+// GetQuoteQtyOk returns a tuple with the QuoteQty field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetQuoteQtyOk() (*string, bool) {
+	if o == nil || common.IsNil(o.QuoteQty) {
+		return nil, false
+	}
+	return o.QuoteQty, true
+}
+
+// HasQuoteQty returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasQuoteQty() bool {
+	if o != nil && !common.IsNil(o.QuoteQty) {
+		return true
+	}
+
+	return false
+}
+
+// SetQuoteQty gets a reference to the given string and assigns it to the QuoteQty field.
+func (o *HistoricalBlockTradesResponseInner) SetQuoteQty(v string) {
+	o.QuoteQty = &v
+}
+
+// GetTime returns the Time field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetTime() int64 {
+	if o == nil || common.IsNil(o.Time) {
+		var ret int64
+		return ret
+	}
+	return *o.Time
+}
+
+// GetTimeOk returns a tuple with the Time field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetTimeOk() (*int64, bool) {
+	if o == nil || common.IsNil(o.Time) {
+		return nil, false
+	}
+	return o.Time, true
+}
+
+// HasTime returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasTime() bool {
+	if o != nil && !common.IsNil(o.Time) {
+		return true
+	}
+
+	return false
+}
+
+// SetTime gets a reference to the given int64 and assigns it to the Time field.
+func (o *HistoricalBlockTradesResponseInner) SetTime(v int64) {
+	o.Time = &v
+}
+
+// GetIsBuyerMaker returns the IsBuyerMaker field value if set, zero value otherwise.
+func (o *HistoricalBlockTradesResponseInner) GetIsBuyerMaker() bool {
+	if o == nil || common.IsNil(o.IsBuyerMaker) {
+		var ret bool
+		return ret
+	}
+	return *o.IsBuyerMaker
+}
+
+// GetIsBuyerMakerOk returns a tuple with the IsBuyerMaker field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *HistoricalBlockTradesResponseInner) GetIsBuyerMakerOk() (*bool, bool) {
+	if o == nil || common.IsNil(o.IsBuyerMaker) {
+		return nil, false
+	}
+	return o.IsBuyerMaker, true
+}
+
+// HasIsBuyerMaker returns a boolean if a field has been set.
+func (o *HistoricalBlockTradesResponseInner) HasIsBuyerMaker() bool {
+	if o != nil && !common.IsNil(o.IsBuyerMaker) {
+		return true
+	}
+
+	return false
+}
+
+// SetIsBuyerMaker gets a reference to the given bool and assigns it to the IsBuyerMaker field.
+func (o *HistoricalBlockTradesResponseInner) SetIsBuyerMaker(v bool) {
+	o.IsBuyerMaker = &v
+}
+
+func (o HistoricalBlockTradesResponseInner) MarshalJSON() ([]byte, error) {
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
+	}
+	return json.Marshal(toSerialize)
+}
+
+func (o HistoricalBlockTradesResponseInner) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !common.IsNil(o.Id) {
+		toSerialize["id"] = o.Id
+	}
+	if !common.IsNil(o.Price) {
+		toSerialize["price"] = o.Price
+	}
+	if !common.IsNil(o.Qty) {
+		toSerialize["qty"] = o.Qty
+	}
+	if !common.IsNil(o.QuoteQty) {
+		toSerialize["quoteQty"] = o.QuoteQty
+	}
+	if !common.IsNil(o.Time) {
+		toSerialize["time"] = o.Time
+	}
+	if !common.IsNil(o.IsBuyerMaker) {
+		toSerialize["isBuyerMaker"] = o.IsBuyerMaker
+	}
+
+	for key, value := range o.AdditionalProperties {
+		toSerialize[key] = value
+	}
+
+	return toSerialize, nil
+}
+
+func (o *HistoricalBlockTradesResponseInner) UnmarshalJSON(data []byte) (err error) {
+	varHistoricalBlockTradesResponseInner := _HistoricalBlockTradesResponseInner{}
+
+	err = json.Unmarshal(data, &varHistoricalBlockTradesResponseInner)
+
+	if err != nil {
+		return err
+	}
+
+	*o = HistoricalBlockTradesResponseInner(varHistoricalBlockTradesResponseInner)
+
+	additionalProperties := make(map[string]interface{})
+
+	if err = json.Unmarshal(data, &additionalProperties); err == nil {
+		delete(additionalProperties, "id")
+		delete(additionalProperties, "price")
+		delete(additionalProperties, "qty")
+		delete(additionalProperties, "quoteQty")
+		delete(additionalProperties, "time")
+		delete(additionalProperties, "isBuyerMaker")
+		o.AdditionalProperties = additionalProperties
+	}
+
+	return err
+}
+
+type NullableHistoricalBlockTradesResponseInner struct {
+	value *HistoricalBlockTradesResponseInner
+	isSet bool
+}
+
+func (v NullableHistoricalBlockTradesResponseInner) Get() *HistoricalBlockTradesResponseInner {
+	return v.value
+}
+
+func (v *NullableHistoricalBlockTradesResponseInner) Set(val *HistoricalBlockTradesResponseInner) {
+	v.value = val
+	v.isSet = true
+}
+
+func (v NullableHistoricalBlockTradesResponseInner) IsSet() bool {
+	return v.isSet
+}
+
+func (v *NullableHistoricalBlockTradesResponseInner) Unset() {
+	v.value = nil
+	v.isSet = false
+}
+
+func NewNullableHistoricalBlockTradesResponseInner(val *HistoricalBlockTradesResponseInner) *NullableHistoricalBlockTradesResponseInner {
+	return &NullableHistoricalBlockTradesResponseInner{value: val, isSet: true}
+}
+
+func (v NullableHistoricalBlockTradesResponseInner) MarshalJSON() ([]byte, error) {
+	return json.Marshal(v.value)
+}
+
+func (v *NullableHistoricalBlockTradesResponseInner) UnmarshalJSON(src []byte) error {
+	v.isSet = true
+	return json.Unmarshal(src, &v.value)
+}