+++ clients/derivativestradingportfoliomargin/CHANGELOG.md
@@ -1,5 +1,27 @@
 ### Changelog
 
+## 1.8.0 - 2026-05-20
+
+### Changed (4)
+
+#### REST API
+
+- Deleted parameter `closePosition`
+  - affected methods:
+    - `newUmAlgoOrder()` (`POST /papi/v1/um/algo/order`)
+- Modified parameter `quantity`:
+  - required: `false` → `true`
+  - affected methods:
+    - `newUmAlgoOrder()` (`POST /papi/v1/um/algo/order`)
+- Modified response for `newUmAlgoOrder()` (`POST /papi/v1/um/algo/order`):
+  - property `closePosition` deleted
+- Modified response for `cancelUmAlgoOrder()` (`DELETE /papi/v1/um/algo/order`):
+  - property `complete` added
+  - property `clientAlgoId` deleted
+  - property `code` deleted
+  - property `msg` deleted
+  - property `algoId` deleted
+
 ## 1.7.0 - 2026-04-20
 
 ### Added (7)